Fix warning.
git-svn-id: https://svn.code.sf.net/p/gxemul/code2/gxemul/trunk@6234 5063f80b-35bb-49d0-95f4-ce7ec8300d17

+++ src/machines/machine_pmax.cc
@@ -876,7 +876,7 @@ abort();
 	    (unsigned char *)&machine->md.pmax->memmap->pagesize, 4096);
 
 	for (unsigned int j=0; j<sizeof(machine->md.pmax->memmap->bitmap); j++)
-		machine->md.pmax->memmap->bitmap[j] = ((int)j * 4096*8 <
+		machine->md.pmax->memmap->bitmap[j] = (j * 4096*8 <
 		    1048576*machine->physical_ram_in_mb)? 0xff : 0x00;
 
 	store_buf(cpu, DEC_MEMMAP_ADDR,
